Time: 37 ms (63.2%), Space: 16.7 MB (44.59%) - LeetHub

+++ 0149-max-points-on-a-line/0149-max-points-on-a-line.cpp
@@ -13,7 +13,7 @@ public:
                 } else if (p2[0] == p1[0]) {
                     vertical++;
                 } else {
-                    double slope = static_cast<double>(p2[1] - p1[1]) / (p2[0] - p1[0]);
+                    double slope = double(p2[1] - p1[1]) / double(p2[0] - p1[0]);
                     count[slope]++;
                     localMax = max(localMax, count[slope] + 1);
                 }
